Edit contents of confirmation email
The upcoming `4.1.1+portage-4.2.3` release will require all existing users to confirm their emails. This commit edits the contents of that confirmation email. "Thank you for registering." makes sense for users that just created an account, but doesn't sound right when read by users with existing accounts. Thus, we are removing the sentence.