Add `config/locale/*` to `.dockerignore`
This commit updates `.dockerignore` to exclude all files in the `config/locale/` directory except for `i18n_placeholders.rb`. The locale files are managed and updated via a shared volume, which is updated periodically through a cron job that executes the `translation:sync` task. This task requires access to the production database to keep the locale files in sync.